requirements.txt
tensorflow-cpu==2.5.0
pandas==1.2.4
numpy==1.19.5
flask==1.1.2
gunicorn==20.1.0
opencv-python==4.5.2.52
Werkzeug==2.0.2